+++ src/phpDocumentor/Plugin/Core/Transformer/Writer/Twig.php
@@ -219,7 +219,10 @@ class Twig extends WriterAbstract implements Routable
             array_unshift($templateFolders, $path);
         }
 
-        $env = new \Twig_Environment(new \Twig_Loader_Filesystem($templateFolders));
+        $env = new \Twig_Environment(
+            new \Twig_Loader_Filesystem($templateFolders),
+            array('cache' => sys_get_temp_dir() . '/phpdoc-twig-cache')
+        );
 
         $this->addPhpDocumentorExtension($project, $transformation, $destination, $env);
         $this->addExtensionsFromTemplateConfiguration($transformation, $project, $env);
@@ -271,21 +274,29 @@ class Twig extends WriterAbstract implements Routable
         ProjectDescriptor $project,
         \Twig_Environment $twigEnvironment
     ) {
-        /** @var \SimpleXMLElement $extension */
-        foreach ((array) $transformation->getParameter('twig-extension', array()) as $extension) {
-            $extension = (string) $extension;
-            if (!class_exists($extension)) {
-                throw new \InvalidArgumentException('Unknown twig extension: ' . $extension);
+        $isDebug = $transformation->getParameter('twig-debug')
+            ? $transformation->getParameter('twig-debug')->getValue()
+            : false;
+        if ($isDebug == 'true') {
+            $twigEnvironment->enableDebug();
+            $twigEnvironment->addExtension(new \Twig_Extension_Debug());
+        }
+
+        /** @var Template\Parameter $extension */
+        foreach ($transformation->getParametersWithKey('twig-extension') as $extension) {
+            $extensionValue = $extension->getValue();
+            if (!class_exists($extensionValue)) {
+                throw new \InvalidArgumentException('Unknown twig extension: ' . $extensionValue);
             }
 
             // to support 'normal' Twig extensions we check the interface to determine what instantiation to do.
             $implementsInterface = in_array(
                 'phpDocumentor\Plugin\Core\Twig\ExtensionInterface',
-                class_implements($extension)
+                class_implements($extensionValue)
             );
 
             $twigEnvironment->addExtension(
-                $implementsInterface ? new $extension($project, $transformation) : new $extension()
+                $implementsInterface ? new $extensionValue($project, $transformation) : new $extensionValue()
             );
         }
     }