+++ CHANGELOG.md
@@ -1,11 +1,44 @@
+2013/09/28: Version 2.1
+-----------------------
+
+```
+ADDED:     New cli option 'log' to tell phpDocumentor the path where to log to
+ADDED:     Verbosity can now be provided in three level, each indicating how much is written in the log
+ADDED:     Support for the XSLCache drop in replacement of ext/xsl
+ADDED:     Locale can now be supplied in the configuration
+ADDED:     German translation for error messages
+ADDED:     Support to insert Markdown files into specific points of the outputted documentation (partials)
+ADDED:     Deprecated tag now registers the version number since when the associated element was deprecated
+ADDED:     Show which traits are consumed by a class
+ADDED:     Link to the traits that are consumed by a class
+ADDED:     Show traits with their methods and properties
+CHANGED:   phpDocumentor now checks for transformation requirements at the start of the application (#148)
+FIXED:     Checkstyle error report only showed errors of the file itself and not subelements (#1046)
+FIXED:     Validation to check for validity of return types
+FIXED:     Validation to check for validity of parameters
+FIXED:     Fatal error when a parent interface is not in the project
+FIXED:     Template:list does not throw an error
+FIXED:     XML output correctly exposes @see, @link, @version
+FIXED:     XML output now does not choke on special characters
+FIXED:     Error code for return type was incorrect
+FIXED:     Functions in the responsive twig now unfold to show complete contents
+FIXED:     Constants in the responsive twig now unfold to show complete contents
+FIXED:     Using an ampersand in the type of an @param no longer causes a fatal (not a recommended practice!)
+REMOVED:   #814; removed ParserAbstract base class for Parser
+REMOVED:   PSR Draft is now moved to its own repo: https://github.com/phpDocumentor/fig-standards/blob/master/proposed/phpdoc.md
+REMOVED:   Automatic generation of log files, the new log option can be used or a configuration option
+```
+
 2013/08/08: Version 2.0.1
 -------------------------
 
+```
 FIXED:      Generated phar files could not be set to executable and ran.
 FIXED:      Missing File-level DocBlocks were not detected.
 FIXED:      Classes and namespace were not generated in Windows.
 FIXED:      Notices were thrown with the Clean template.
 FIXED:      In windows were path calculated incorrectly.
+```
 
 2013/08/03: Version 2.0.0
 -------------------------